moves files to related folders
