+++ src/ethereum/ethereumProofPointIssueResult.ts
@@ -1,0 +1,9 @@
+import { ProofPoint } from "../proofPoint";
+import { ProofPointId } from "../proofPointId";
+import { EthereumTransactionHash } from "./ethereumTransactionHash";
+
+export interface EthereumProofPointIssueResult {
+  proofPointId: ProofPointId;
+  transactionHash: EthereumTransactionHash;
+  proofPointObject: ProofPoint;
+}
