+++ src/ipfs/ipfsStorageProviderSettings.ts
@@ -1,0 +1,5 @@
+export interface IpfsStorageProviderSettings {
+  host: string;
+  port: number;
+  protocol?: string;
+}
