+++ src/proofPointResolver.ts
@@ -1,6 +1,5 @@
 import { ProofPoint } from "./proofPoint";
-import { ProofPointId, ProofPointIdType } from "./proofPointId";
-import { StorageProvider } from "./storage";
+import { ProofPointId } from "./proofPointId";
 
 /**
  * Proof point resolver
@@ -13,27 +12,4 @@ export interface ProofPointResolver {
    * @returns The @ProofPoint corresponding to the given ID.
    */
   resolve(id: ProofPointId): Promise<ProofPoint>;
-}
-
-/**
- * Ipfs proof point resolver
- * A @ProofPointResolver capable of handling @ProofPointId of type Ipfs
- */
-export class IpfsProofPointResolver {
-  private _storage: StorageProvider;
-
-  public constructor(storage: StorageProvider) {
-    this._storage = storage;
-  }
-
-  public async resolve(id: ProofPointId): Promise<ProofPoint> {
-    if (id.getType() !== ProofPointIdType.Ipfs) {
-      throw `Unresolvable ID type: ${id.getType()}`;
-    }
-
-    const ipfsAddress = id.toString();
-    const storedData = await this._storage.get(ipfsAddress);
-    const proofPoint = JSON.parse(storedData.data);
-    return proofPoint;
-  }
 }
